arm/arm64: KVM: Feed initialized memory to MMIO accesses
On an MMIO access, we always copy the on-stack buffer info
the shared "run" structure, even if this is a read access.
This ends up leaking up to 8 bytes of uninitialized memory
into userspace.

An obvious fix for this one is to only perform the copy if
this is an actual write.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>