arm64: KVM: vgic-v3: Do not save ICH_AP0Rn_EL2 for GICv2 emulation
The GICv3 specification mandates that ICH_AP0Rn_EL2 are set to
zero when running guests that use the memory mapped registers.

This is fine, as we initialize all ICH_AP0Rn_EL2 registers to
zero, and restore them on entry. But it also means that we
do not need to save these registers on exit. Profit!

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>