KVM: arm/arm64: vgic-new: Change MMIO handlers
Change type of read and write functions to not take/return a void * but
instead take and return an unsigned long.  There is no need to pass the
iodev to each handler and the redistributor vcpu can be resolved at the
caller.

Also, get rid of header definitions of functions not used outside
vgic_mmio.c

Note: This is not actually expected to work on anything else than
LE guest on LE host setups so far.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>