KVM: arm/arm64: vgic-new: Rework uaccess functions and remove unused
Some GICv3 functions are unused so far and not knowing how we're going
to finalize the API for exporting this state to userspace, let's just
wait with adding in the info until when we actually implement uaccess
for GICv3.

For the GICv2 functions, we need to do the endianness fiddling and pack
the data in the right way like we used to in the old vgic for this to
work on BE systems.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>