arm/arm64: KVM: Enforce HYP read-only mapping of the kernel's rodata section
In order to be able to use C code in HYP, we're now mapping the kernel's
rodata in HYP. It works absolutely fine, except that we're mapping it RW,
which is not what it should be.

Add a new HYP_PAGE_RO protection, and update create_hyp_mappings() to take
a RO flag, which gets used when mapping the rodata section.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>