KVM: arm/arm64: vgic: Reject GICV mappings that don't match host alignment
We currently allow the GICV region to be mapped at any 4kB offset.
This is always fine with 4kB pages on the host, but leads to odd
results if the host uses a bigger page size (16kB or 64kB), and
the GICV region not aligned on a page boundary (a typical
example is to have the two 4kB regions stradling 2 64kB pages).

In this case, userspace has 1 chance in 16 to place the GICV
region at the right sub-page offset (it will likely miss it),
leading to a guest that cannot access its GIC CPU interface
(because it has been provided with the wrong address).

This patch checks this alignment constraints and returns an
error to userspace instead of continuing blindly.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>