irqchip: gic: Drop useless locking when handling cascaded interrupt
When handling a cascaded interrupt, we take the GIC lock, which
is odd and completely useless since the interrupt is handled on
a single CPU. Also, in the case where there would be multiple
cascaded GICs, there is strictly no need to serialize them.
Also, this lock only covers the logical to physical CPU map,
which is neither written nor read.

Given the above, dropping the lock seems to be a reasonable
course of action.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>