irqchip/gic-v3-its: Support VPE doorbell invalidation even when !DirectLPI
When we don't have the DirectLPI feature, we must work around the
architecture shortcomings to be able to perform the required
invalidation.

For this, we create a fake device whose sole purpose is to
provide a way to issue a map/inv/unmap sequence (and the corresponding
sync operations). That's 6 commands and a full serialization point
to be able to do this.

You just have to hope the hypervisor won't do that too often...

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>