irqdomain: Tidy-up inconsistent checks in virq_debug_show_one
We check for the irq_data being NULL long after having dereferenced
it, and while there is no chance of it being NULL, let's put the
check at its logical spot. This can then be used for more advanced
checks (such as the depth of the irq_data and domain hierarchies,
which should always be the same).

Reported-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>