arm64: perf: Use v8.0 behaviour when filtering for a guest
When KVM installs a perf event on behalf of a guest, special care
must be taken to count events happening at EL1 (because that's
where the guest kernel is running) even if the host kernel runs
at EL2 (which is the case with VHE on a v8.1+ CPU).

This relies on the exclude_host flag set by KVM's PMU emulation.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>