PCI/MSI: Move MSI domain discovery to the allocation point
Now that we're able to discover the MSI domain in a more "lazy" way,
let's stop populating the msi_domain field in the probing code, and
instead assign it when we actually need it, in the PCI/MSI code.

For this, we introduce a new helper (pci_get_msi_domain), which does
the heavy lifting of walking the PCI hierarchy and return the domain
(or NULL if none can be found).

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>