PCI: layerscape: Cleanup ls_pcie_msi_host_init
ls_pcie_msi_host_init() has some irrelevant comments, check for
properties that may or may not be present ("msi-parent"), and
fails to account for the ones that should be there, such as
"msi-map". Overall, this is a waste of space, as only the
end-point is truly aware of whether it has an MSI domain or
not.

So let's reduce thus function to what it actually is, a simple
placeholder that prevents the generic DWC driver to instanciate
its own MSI controller.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>